GH Actions: report linting violations in the PR
The cs2pr tool will allow to display the results from an action run in checkstyle format in-line in the PR code view, which should improve usability of the workflow results.

This implements this for the results of running Parallel Lint over its own code.

This also implicitly acts as an integration test for the `checkstyle` output and the cs2pr integration.

Ref: https://github.com/staabm/annotate-pull-request-from-checkstyle